requirements.txt
# Core FastAPI dependencies
fastapi[all]==0.109.0
uvicorn[standard]==0.27.0

# Frontend
streamlit==1.31.0

# LLM Integration
google-generativeai==0.3.2

# Docker SDK
docker==7.0.0

# Configuration & Environment
python-dotenv==1.0.0

# HTTP requests (for Streamlit frontend)
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2  # Pinned for TestClient compatibility  # For TestClient

# Code Quality
black==23.12.1
flake8==6.1.0
mypy==1.7.1
